prevent re-render on drawer open and exit dialog
